Created reinstantiation middleware Bumped version
