+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@caal-15/winston3-logstash-transport",
-  "version": "1.0.3",
+  "version": "1.0.4",
   "description": "A winston@3 replacement for both winston-logstash and winston-logstash-udp to facilitate either TCP or UDP traffic to logstash",
   "main": "index.js",
   "scripts": {
